requirement.txt
customtkinter
Pillow
numpy
PyPDF2
python-docx
